requirements.txt
kanboard
python-dotenv
pytest
pyyaml
requests
